Really long commit, should have broken this up
Created the turtle class with many of its attributes. Still need to add
angle and add methods for adjusting rotation of the turtle.

Added new button in GUI to change color of pane as well as experimented
with moving the turtle around with some test methods. Added drawing
lines and next is to add changing the pen color through a button.